backend/requirements.txt
# FastAPI Backend
fastapi
uvicorn[standard]
python-multipart

# Config & Environment
pydantic
pydantic-settings
python-dotenv

# PDF Processing
PyMuPDF

# RAG + LLM
langchain
langchain-community

# Embeddings (IMPORTANT ✅)
sentence-transformers

# Vector Database
pinecone-client

# LLM (Groq)
groq

# Requests
requests

# Utilities
tiktoken
numpy

# Testing 
pytest